requirements.txt
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1
aiogram~=3.6.0